Move Resolution out of CameraView
Add long-press to show/hide controls
